fix abstract type merge.
